ADK-Modules/requirements.txt
google-adk
python-dotenv
yfinance
psutil
litellm
google-generativeai